Add communication agent-improc, improc-agent, agent-simulation

+++ AgentSoftware/MappingAgentSoftware/src/Point.java
@@ -1,17 +1,17 @@
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
